CTimeGAN/requirements.txt
numpy~=1.21.2
tensorflow-gpu~=2.8.0
setuptools~=58.0.4
pandas~=1.2.5
tqdm~=4.62.2